+++ ClientModels/Runitime/Scripts/Models/CMTemplateObj.cs
@@ -27,7 +27,11 @@ namespace Reflectis.SDK.ClientModels
             public int id;
 
             // SceneObj: Downloaded3DModel---ColorPicker
-            public Color modelColor = Color.white;
+            public string modelColor = "";
+
+            // SceneObj: Downloaded3DModel---Explodable
+            public float modelExplosionValue;
+            public int modelExplosionType;
 
             // SceneObj: Downloaded3DPickables
             public string addressableString;