+++ InteractionNew/Runtime/Scripts/ContextualMenu/ContextualMenuSystem.cs
@@ -21,10 +21,6 @@ namespace Reflectis.SDK.InteractionNew
         protected bool esclusiveContextualMenus = false;
         [SerializeField] protected InputActionReference contextualMenuInputActionRef;
 
-        [Header("Scriptable actions")]
-        [SerializeField] private List<AwaitableScriptableAction> onHoverEnterActions = new();
-        [SerializeField] private List<AwaitableScriptableAction> onHoverExitActions = new();
-
         protected ContextualMenuController contextualMenu;
         protected ContextualMenuManageable selectedInteractable;
 
@@ -36,11 +32,7 @@ namespace Reflectis.SDK.InteractionNew
         public float ShowToastTime { get => showTime; private set => showTime = value; }
         public float HideToastTime { get => hideTime; private set => hideTime = value; }
 
-        public List<AwaitableScriptableAction> OnHoverEnterActions => onHoverEnterActions;
-        public List<AwaitableScriptableAction> OnHoverExitActions => onHoverExitActions;
-
-
-        public override void Init()
+        public override Task Init()
         {
             if (createMenuOnInit)
             {
@@ -52,6 +44,7 @@ namespace Reflectis.SDK.InteractionNew
             contextualMenuInputActionRef.action.started += OnMenuActivate;
             contextualMenuInputActionRef.action.performed += OnMenuActivate;
             contextualMenuInputActionRef.action.canceled += OnMenuActivate;
+            return base.Init();
         }
 
         private void OnDestroy()
@@ -133,21 +126,6 @@ namespace Reflectis.SDK.InteractionNew
                 await contextualMenu.Hide();
                 contextualMenu.Unsetup();
             }
-        }
-
-        public void ShowPreviewContextualMenu(ContextualMenuManageable manageable)
-        {
-            if (customContextualMenuControllersCache.TryGetValue(manageable.ContextualMenuType, out contextualMenu))
-            {
-                contextualMenu.Setup(manageable.ContextualMenuOptions);
-                contextualMenu.ShowPreview();
-            }
-        }
-
-        public void HidePreviewContextualMenu()
-        {
-            contextualMenu.HidePreview();
-            contextualMenu.Unsetup();
         }
 
         #endregion