+++ InteractionNew/Runtime/Scripts/Manipulation/ManipulationSystemBase.cs
@@ -1,7 +1,5 @@
 using Reflectis.SDK.Core;
 
-using System.Collections.Generic;
-
 using UnityEngine;
 
 namespace Reflectis.SDK.InteractionNew
@@ -11,17 +9,12 @@ namespace Reflectis.SDK.InteractionNew
         [SerializeField] protected GameObject scalablePointFacePrefab;
         [SerializeField] protected Vector2 targetScaleFaces = new Vector3(0.15f, 0.15f, 0.15f);
 
-        [Header("Scriptable actions")]
-        [SerializeField] protected List<AwaitableScriptableAction> onHoverEnterActions = new();
-        [SerializeField] protected List<AwaitableScriptableAction> onHoverExitActions = new();
-
         public GameObject ScalablePointFacePrefab => scalablePointFacePrefab;
 
-        public List<AwaitableScriptableAction> OnHoverEnterActions => onHoverEnterActions;
-        public List<AwaitableScriptableAction> OnHoverExitActions => onHoverExitActions;
+        public abstract void OnManipulableHoverEnter(IInteractable interactable);
+        public abstract void OnManipulableHoverExit(IInteractable interactable);
 
         public abstract InteractableBehaviourBase SetupInteractableBehaviour(GameObject obj);
 
-        public override void Init() { }
     }
 }