+++ InteractionNew/Runtime/Scripts/VisualScripting/HoverExitEventUnit.cs
@@ -1,0 +1,13 @@
+using Unity.VisualScripting;
+
+namespace Reflectis.SDK.InteractionNew
+{
+    [UnitTitle("Reflectis Generic Interactable: On Hover Exit")]
+    [UnitSurtitle("Generic Interactable")]
+    [UnitShortTitle("On Hover Exit")]
+    [UnitCategory("Events\\Reflectis")]
+    public class HoverExitEventUnit : GenericInteractableEventUnit
+    {
+
+    }
+}