+++ InteractionNew/Runtime/Scripts/VisualScripting/SelectEnterEventUnit.cs
@@ -1,0 +1,13 @@
+using Unity.VisualScripting;
+
+namespace Reflectis.SDK.InteractionNew
+{
+    [UnitTitle("Reflectis Generic Interactable: On Select Enter")]
+    [UnitSurtitle("Generic Interactable")]
+    [UnitShortTitle("On Select Enter")]
+    [UnitCategory("Events\\Reflectis")]
+    public class SelectEnterEventUnit : GenericInteractableEventUnit
+    {
+
+    }
+}