+++ Utilities/Runtime/Scripts/IApplicationManager.cs
@@ -1,0 +1,13 @@
+namespace Reflectis.SDK.Utilities
+{
+    public interface IApplicationManager
+    {
+        public static IApplicationManager Instance { get; protected set; }
+
+        public void QuitApplication();
+
+        public void ErasePlayerSessionData();
+
+        public void HideEverything();
+    }
+}